% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseRealData.R
\name{analyse.realdata}
\alias{analyse.realdata}
\title{Compute p-value for the overall treatment effect and for the 
interaction effect between the treatment and the predicted sensitivity status, for real data.}
\usage{
analyse.realdata(datalist, sig = 0.05, group.prop.sig = 0.2,
  method = c("cvrs", "cvasd"), eta = NULL, R = NULL, G = NULL,
  seed = NULL, plotrs = F)
}
\arguments{
\item{datalist}{A list made up of 3 data frames (patients and covar) and a vector of binary responses, for real data (see \code{realdata} object): patients (a data frame with patients information), covar (a data frame with the covariates), response (a vector of responses).}

\item{sig}{An overall significance level for the adaptive design.}

\item{group.prop.sig}{Proportion of significance level for the sensitive group test.}

\item{method}{"cvasd" (for the cross-validated adaptive signature design) or "cvrs" (for the cross-validated risk scores design).}

\item{eta}{A significance level for covariate-wise logistic regression for the "cvasd" method (double, or vector of doubles).}

\item{R}{A threshold of the odds ratio for the "cvasd" method (double, or vector of doubles).}

\item{G}{A threshold for the number of covariates for the "cvasd" method (integer, or vector of integers).}

\item{seed}{A  seed for the random number generator.}

\item{plotrs}{An indicator whether to plot the risk scores for the "cvrs" method (default = FALSE).}
}
\value{
An object of class \code{"rapids"}.

patients: A data frame with patients inormation.

pval.overall: P-value for the overall test for the treatment effect.

stat.group: Statistic for the interaction effect between the treatment and the predicted sensitivity status.

pval.group: P-value for the interaction effect between the treatment and the predicted sensitivity status.

sens.pred: Predicted sensitivity status (rows = patients, columns = simulations).

estimate.rr: Estimated response rate in the sensitive group on the treatment arm.

cvrs: A matrix of the risk scores (rows = patients, columns = simulations), for "cvrs" method only.

eta,R,G: Significance level, threshold of the odds ratio and threshold for the number of covariates for covariate-wise logistic regression, for"cvasd" method only.
}
\description{
The sensitive group is identified according to the input method ("cvasd" for the cross-validated adaptive signature design or "cvrs" for the cross-validated risk scores design),
significance level of the test for the sensitive group is 
determined by the input proportion of the significance  
level for the sensitive group (group.prop.sig) out of the overall significance (sig).
The p-value for the overall treatment effect is computed using logistic regression with treatment allocation as a predictor.
The p-value for the interaction effect between the treatment and the predicted sensitivity status is computed using logistic regression with the treatment effect, the effect for the predicted sensitivity status and the interaction effect.
}
\examples{
#"cvrs" method
sig = 0.05
group.prop.sig = 0.2
method = "cvrs"
eta = NULL
R = NULL
G = NULL
seed = 123
plotrs = T
realres.cvrs = analyse.realdata(realdata, sig, group.prop.sig, method, eta, R, G, seed, plotrs)

#"cvasd" method
sig = 0.05
group.prop.sig = 0.2
method = "cvasd"
eta = c(0.01, 0.02, 0.03)
R = c(2.5, 2, 1.5)
G = c(3,2,1)
seed = 123
plotrs = F
realres.cvasd = analyse.realdata(realdata, sig, group.prop.sig, method, eta, R, G, seed, plotrs)
}
\seealso{
\code{\link{analyse.simdata}}, \code{\link{simulate.data}} and \code{\link{cvrs.plot}} functions; \code{\link{print}} and \code{\link{plot}} methods.
}
\author{
Svetlana Cherlin, James Wason
}
