% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseRealData.R
\name{analyse.realdata2}
\alias{analyse.realdata2}
\title{Compute p-value for the overall treatment effect and for the
interaction effect between the treatment and the predicted sensitivity status, for real data.}
\usage{
analyse.realdata2(
  datalist,
  nclust,
  sig = 0.05,
  group.prop.sig = 0.2,
  seed = NULL,
  plotrs = F
)
}
\arguments{
\item{datalist}{A list made up of 3 data frames (patients and covar) and a vector of binary responses, for real data (see \code{realdata} object): patients (a data frame with patients information), covar (a data frame with the covariates), response (a vector of responses).}

\item{nclust}{Number of clusters}

\item{sig}{An overall significance level for the adaptive design.}

\item{group.prop.sig}{Proportion of significance level for the sensitive group test.}

\item{seed}{A  seed for the random number generator.}

\item{plotrs}{An indicator whether to plot the risk scores (default = FALSE).}
}
\value{
An object of class \code{"rapids"}.

patients: A data frame with patients inormation.

pval.resp: P-value for the overall test for the treatment effect w.r.t. response 1

pval.resp2: P-value for the overall test for the treatment effect w.r.t. response 2

stat.resp.group: Statistic for the interaction effect between the treatment  and the predicted cluster (w.r.t. response ), a vector of length nclust

pval.resp.group: P-value for the interaction effect between the treatment and the and the predicted cluster (w.r.t. response 2), a vector of length nclust

stat.resp2.group: Statistic for the interaction effect between the treatment and the predicted cluster (w.r.t. response), a vector of length nclust

pval.resp2.group: P-value for the interaction effect between the treatment and the predicted cluster (w.r.t. response 2), a vector of length nclust

estimate.rr: Estimated rate of responsein the sensitive group on the treatment arm.

estimate.rr2: Estimated rate of response 2 in the sensitive group on the treatment arm.

cvrs: A matrix of the risk scores (rows = patients, columns = simulations).

cvrs2: A matrix of the risk scores 2(rows = patients, columns = simulations).

cluster.pred: Predicted clusters
}
\description{
significance level of the test for the sensitive group is
determined by the input proportion of the significance
level for the sensitive group (group.prop.sig) out of the overall significance (sig).
The p-value for the overall treatment effect is computed using bivariate logistic regression with treatment allocation as a predictor.
The p-value for the interaction effect between the treatment and the predicted sensitivity status is computed using bivariate logistic regression with the treatment effect, the effect for the predicted sensitivity status and the interaction effect.
}
\examples{
sig = 0.05
nclust = 4
group.prop.sig = 0.2
seed = 123
plotrs = T
realres.cvrs2 = analyse.realdata2(realdata2, nclust, sig, group.prop.sig, seed, plotrs)

}
\seealso{
\code{\link{analyse.simdata2}}, \code{\link{simulate.data2}} and \code{\link{cvrs.plot}} functions; \code{\link{print}} and \code{\link{plot}} methods.
}
\author{
Svetlana Cherlin, James Wason
}
