% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{cvrs.plot}
\alias{cvrs.plot}
\title{Plot the risk scores.}
\usage{
cvrs.plot(cvrs, sens.pred)
}
\arguments{
\item{cvrs}{A data frame of the  risk scores.}

\item{sens.pred}{A data frame of the predicted sensitivity statuses.}
}
\description{
The function plots the risk scores.
}
\details{
A plot of the risk scores is produced and saved into a "CVRSplot.png" file.
}
\examples{
#Analyse simulated data and plot the risk scores
#"cvrs" method
data(simdata)
sig = 0.05
group.prop.sig = 0.2
seed = 123
plotrs = T
method = "cvrs"
eta = NULL
R = NULL
G = NULL
simres.cvrs = analyse.simdata(simdata, sig, group.prop.sig, method, eta, R, G, seed, plotrs)
 #Plot the risk scores
cvrs.plot(simres.cvrs$cvrs, simres.cvrs$sens.pred)
}
\seealso{
\code{\link{analyse.simdata}} and \code{\link{analyse.realdata}} functions; \code{\link{plot}} and \code{\link{print}} methods.
}
\author{
Svetlana Cherlin, James Wason
}
