% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{cvrs2.plot}
\alias{cvrs2.plot}
\title{Plot the risk scores for two outcomes}
\usage{
cvrs2.plot(cvrs, cvrs2, cluster.pred, cluster.true = NULL, sim = FALSE)
}
\arguments{
\item{cvrs}{A data frame of the  risk scores 1}

\item{cvrs2}{A data frame of the  risk scores 2}

\item{cluster.pred}{True clusters}

\item{sim}{An indicator of whether the data is simulated or real}

\item{sens.pred}{A data frame of the predicted sensitivity statuses}
}
\value{
A ggplot object
}
\description{
The function plots the risk scores
}
\details{
A plot of the risk scores is produced and saved into a ggplot object
}
\examples{
#Analyse simulated data and plot the risk scores
data(simdata)
sig = 0.05
group.prop.sig = 0.2
seed = 123
plotrs = T
simres.cvrs2 = analyse.simdata2(simdata2, sig, group.prop.sig, seed, plotrs)
 #Plot the risk scores
cvrs2.plot(simres.cvrs2$cvrs, simres.cvrs2$cvrs2, simres.cvrs2$cluster.pred, simdata2$patients$cluster.true, TRUE)
}
\seealso{
\code{\link{analyse.simdata2}} and \code{\link{analyse.realdata2}} functions; \code{\link{plot}} and \code{\link{print}} methods.
}
\author{
Svetlana Cherlin, James Wason
}
