% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get.tuning.param}
\alias{get.tuning.param}
\title{Get tuning parameters for CVASD by nested cross-validation.}
\usage{
get.tuning.param(patients.train, covar.train, y.train, eta, R, G, seed)
}
\arguments{
\item{patients}{- a data frame of patients inormation 
covar - a data frame of covariates
y - a vector of responses 
eta - significance level for covariate-wise logistic regression for "cvasd" method (double of vector of doubles)
R - a threshold of the odds ratio for "cvasd" method (double of vector of doubles)
G - a threshold for the number of covar for "cvasd" method (integer of vector of integers)
seed - a seed for random number generator}
}
\value{
A data frame with the optimal eta, R, and G
}
\description{
An internal function for tuning the parameters.
In one inner fold only, compute p-value for the test of treatment effect comparison in the 
selected sensitive group, for each combination of (eta, R, G). Output the (eta, R, G) combination that
corresponds to the smallest p-value.
}
\author{
Svetlana Cherlin, James Wason
}
