% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{permutation.test}
\alias{permutation.test}
\title{Perform permutation test for the real data (1 outcome)}
\usage{
permutation.test(realdata, res, method, reps)
}
\arguments{
\item{realdata}{Input data (see \code{realdata} object).}

\item{res}{An object of class "rapids" containing the results for the design for the original data.}

\item{method}{An input method ("cvasd" or "cvrs").}

\item{reps}{Number of permutations.}
}
\value{
A list of 5:

stat0.group:  Statistic for the interaction effect between the treatment and the sensitivity status, for the original data.

treat: A matrix of treatment allocations, one column per permutation run.

sens.pred: A matrix of sesnsitivy statuses, one column per permutation run.

stat.group:  A matrix of permutation statistics, one column per permutation run.

ppval.group: P-value from the permutation test.
}
\description{
For each permutation run, a test statistic for the interaction effect between the treatment and the predicted sensitivity status is computed.
A permutation p-value is computed as the proportion of simulations where the test statistic is larger than the test statistic for the original data.
}
\details{
Perform permutation test for the real data.
}
\examples{
#Analyse data with "cvrs" method and perform permutation test
data(realdata)
sig = 0.05
group.prop.sig = 0.2
method = "cvrs"
seed = 123
plotrs = T
eta = NULL
R = NULL
G = NULL 
realres.cvrs = analyse.realdata(realdata, sig, group.prop.sig, method, eta, R, G, seed, plotrs)
#Permutation test
reps = 10
res.perm = permutation.test(realdata, realres.cvrs, method, reps)
}
\seealso{
\code{\link{analyse.realdata}} function.
}
\author{
Svetlana Cherlin, James Wason
}
