% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{plot.rapids}
\alias{plot.rapids}
\title{Plot risk scores, sensitivity and specificity from a "rapids" object.}
\usage{
\method{plot}{rapids}(x, ...)
}
\arguments{
\item{x}{An object of class "rapids"}

\item{...}{Other parameters to plot}
}
\description{
Plot risk scores, sensitivity and specificity from a "rapids" object.
}
\details{
For simulated data, the following is plotted: ROC plots for predicting the sensitivity status, boxplots for the AUC, sensitivity and specificity, risk scores for the "cvrs" method (the risk scores from all of the simulation runs are shown on the same plot).

For real data, risk scores are plotted for the "cvrs" method. For real data  analysed with the "cvasd" method, a message "nothing to plot" is printed.
}
\examples{
#Analyse simulated data and plot the risk scores
#"cvrs" method
data(simdata)
sig = 0.05
group.prop.sig = 0.2
seed = 123
plotrs = T
method = "cvrs"
eta = NULL
R = NULL
G = NULL
simres.cvrs = analyse.simdata(simdata, sig, group.prop.sig, method, eta, R, G, seed, plotrs)
#Plot the results
plot(simres.cvrs)
}
\seealso{
\code{\link{cvrs.plot}} function and \code{\link{print}} method.
}
\author{
Svetlana Cherlin, James Wason
}
