% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{print.rapids}
\alias{print.rapids}
\title{Print an object of class "rapids".}
\usage{
\method{print}{rapids}(x, ...)
}
\arguments{
\item{x}{An object of class "rapids"}
}
\description{
Print an object of class "rapids".
}
\details{
For simulated data, the following operating characteristics are printed: (i) the power for the overall test; (ii) the power for the sensitive group test; (iii) the power for the adaptive design; (iv) mean sensitivity and specificity of identifying the sensitive group; (v) mean response rate in the sensitive group on the treatment arm. For real data, the following operating characteristics are printed: (i) p-value for the overall test; (ii) p-value for the  sensitivity group test; (iii) estimated response rate in the sensitive group on the treatment arm.
}
\examples{
#Analyse simulated data and print the results
#"cvrs" method
data(simdata)
sig = 0.05
group.prop.sig = 0.2
seed = 123
plotrs = T
method = "cvrs"
eta = NULL
R = NULL
G = NULL
simres.cvrs = analyse.simdata(simdata, sig, group.prop.sig, method, eta, R, G, seed, plotrs)
#Print the results
print(simres.cvrs)
}
\seealso{
\code{\link{cvrs.plot}} function and \code{\link{print}} method.
}
\author{
Svetlana Cherlin, James Wason
}
