% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sens.cvasd}
\alias{sens.cvasd}
\title{Get subgroup of sensitive patiens according to the "cvasd" method.}
\usage{
sens.cvasd(patients, covar, y, eta.in, R.in, G.in, seed)
}
\arguments{
\item{patients}{- a data frame of patients inormation 
covar - a data frame of covariates
y - a vector of responses 
eta.in - significance level for covariate-wise logistic regression (double or a vector of doubles)
R.in - a threshold of the odds ratio (double or a vector of doubles)
G.in - a threshold for the number of covariates (integer or a vector of integers)
seed - a seed for random number generator}
}
\value{
A list of 3 :
         psens - sensitivity of identifying the sensitive group, one value per simulation run 
         pspec - specificity of identifying the sensitive group, one value per simulation run 
         sens.pred - predicted sensitivity status (rows = patienst, columns = simulations)
}
\description{
For a vector of binary responses, get sensitive patients by CVASD
according to the input tuning set (eta.in, R.in, G.in).      
In the training subset, fit logistic regression for each covariate.          
In the testing subset, compute the new vs control arm odds ratio for 
covariates that have treatment-covariate interaction significant at a level eta.in
in the training subset. A patient is sensitive if the odds ration > R.in for at least G.in covariates.
If the length of each parameter in the tuning set (eta.in, R.in and G.in) > 1 then the tuning
set is found by a nested cross-validation where only one inner fold is used (get.tuning.param function).
}
\author{
Svetlana Cherlin, James Wason
}
