% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sens.cvrs}
\alias{sens.cvrs}
\title{Get subgroup of sensitive patiens according to the "cvrs" method.}
\usage{
sens.cvrs(patients, covar, y, seed)
}
\arguments{
\item{patients}{- a data frame with patients inormation 
covar - a data frame with covariates
y - a vector of responses 
seed - a seed for random number generator}
}
\value{
A list of 4 :
         psens - sensitivity of identifying the sensitive group, one value per simulation run 
         pspec - specificity of identifying the sensitive group, one value per simulation run 
         sens.pred - predicted sensitivity status (rows = patienst, columns = simulations)
         cvrs - a matrix of the risk scores (rows = patients, columns = simulations).
}
\description{
For one vector of binary responses, get sensitive patients by risk scores and k-means.      
In the testing subsets, the risk scores are computed based on the treatment-covariate interaction
effects from the training subsets.
The risk scores are divided into 2 clusters by a k-means clustering (within each fold).
}
\author{
Svetlana Cherlin, James Wason
}
